bye camino, hello mrtrix

+++ connectome_scripts/f_functional_connectivity.m
@@ -234,6 +234,8 @@ if flags.EPI.SpinEchoUnwarp==1
         warning('%s',paths.EPI.SEFM,' does not exist. Field map correction must be skipped.')
         return
     else
+        fileInAP = fullfile(paths.EPI.SEFM,'AP.nii.gz');
+        fileInPA = fullfile(paths.EPI.SEFM,'PA.nii.gz');
         if exist(paths.EPI.APdcm,'dir') && exist(paths.EPI.PAdcm,'dir')
             fileNiiAP= 'AP';
             sentence=sprintf('rm -fr %s/%s.nii*',paths.EPI.SEFM,fileNiiAP);
@@ -251,18 +253,15 @@ if flags.EPI.SpinEchoUnwarp==1
             
             sentence=sprintf('gzip -f %s/AP.nii %s/PA.nii',paths.EPI.SEFM,paths.EPI.SEFM);
             [~,result] = system(sentence); % gzip fieldmap volumes
-            
+        elseif exist(fileInAP,'file') && exist(fileInPA,'file')
             %Concatenate the AP then PA into single 4D image
             fileOut = fullfile(paths.EPI.SEFM,'sefield.nii.gz');
             if exist(fileOut,'file')
                 sentence=sprintf('rm -fr %s',fileOut);
                 [~,result] = system(sentence); % if it already exists; remove it.
             end
-            fileInAP = fullfile(paths.EPI.SEFM,'AP.nii.gz');
-            fileInPA = fullfile(paths.EPI.SEFM,'PA.nii.gz');
             sentence = sprintf('%s/fslmerge -tr %s %s %s %f',paths.FSL,fileOut,fileInAP,fileInPA,params.EPI.TR);
             [~,result]=system(sentence);
-            
             
             % Generate an acqparams text file based on number of field maps.
             configs.EPI.SEreadOutTime = get_readout(paths,configs.name.dcmFiles);
@@ -459,7 +458,7 @@ if flags.EPI.SpinEchoUnwarp==1
             end
             
         else
-            warning('UNWARP DICOMS folders do exist. Field Map correction failed.')
+            warning('UNWARP DICOMS folders or nii images do not exist. Field Map correction failed.')
             return
         end
     end
@@ -898,8 +897,7 @@ if flags.EPI.DemeanDetrend == 1
             disp(i/sizeX)
         end
     end
-    %fileOut2 = fullfile(paths.EPI.dir,'AROMA','6_epi.nii.gz');
-    fileOut2 = fullfile(paths.EPI.dir,'6_epi.nii.gz');
+    fileOut2 = fullfile(paths.EPI.dir,'AROMA','6_epi.nii.gz');
     MRIwrite(resting,fileOut2,'double');
     sentence=sprintf('%s/fslmaths %s -mas %s %s',paths.FSL,fileOut2,fileOut,fileOut2);
     [~,result]=system(sentence);
@@ -910,10 +908,8 @@ if flags.EPI.Regressors == 1
     disp('---------------------------')
     disp('7. Physiological Regressors')
     disp('---------------------------')
-    %if exist(fullfile(paths.EPI.dir,'AROMA','6_epi.nii.gz'),'file')
-        %fileIn = fullfile(paths.EPI.dir,'AROMA','6_epi.nii.gz');
-    if exist(fullfile(paths.EPI.dir,'6_epi.nii.gz'),'file')
-        fileIn = fullfile(paths.EPI.dir,'6_epi.nii.gz');
+    if exist(fullfile(paths.EPI.dir,'AROMA','6_epi.nii.gz'),'file')
+        fileIn = fullfile(paths.EPI.dir,'AROMA','6_epi.nii.gz');
         resting = MRIread(fileIn);
         [sizeX,sizeY,sizeZ,numTimePoints] = size(resting.vol);
 
@@ -931,8 +927,7 @@ if flags.EPI.Regressors == 1
     GMnumVoxels = nnz(volGM.vol);
     GMmask = logical(repmat(volGM.vol,[1,1,1,numTimePoints]));
     GMts = reshape(resting.vol(GMmask),[GMnumVoxels,numTimePoints]);
-    %save(fullfile(paths.EPI.dir,'AROMA/7_dataGM.mat'),'GMmask','GMts')
-    save(fullfile(paths.EPI.dir,'GSreg_yes/7_dataGM.mat'),'GMmask','GMts')
+    save(fullfile(paths.EPI.dir,'AROMA/7_dataGM.mat'),'GMmask','GMts')
 %-------------------------------------------------------------------------%
     % CSFvent time-series and PCA
     CSFnumVoxels = nnz(volCSFvent.vol);
@@ -942,11 +937,9 @@ if flags.EPI.Regressors == 1
     CSFderiv = [0,diff(CSFavg)];
     if configs.EPI.aCompCor == 1
         [CSFpca,CSFvar] = get_pca(CSFts',configs.EPI.numCompPCA);
-        %save(fullfile(paths.EPI.dir,'AROMA/7_dataCSF.mat'),'CSFpca','CSFvar','CSFmask','CSFts','CSFavg','CSFderiv');
-        save(fullfile(paths.EPI.dir,'GSreg_yes/7_dataCSF.mat'),'CSFpca','CSFvar','CSFmask','CSFts','CSFavg','CSFderiv');
+        save(fullfile(paths.EPI.dir,'AROMA/7_dataCSF.mat'),'CSFpca','CSFvar','CSFmask','CSFts','CSFavg','CSFderiv');
     else
-        %save(fullfile(paths.EPI.dir,'AROMA/7_dataCSF.mat'),'CSFmask','CSFts','CSFavg','CSFderiv');
-        save(fullfile(paths.EPI.dir,'GSreg_yes/7_dataCSF.mat'),'CSFmask','CSFts','CSFavg','CSFderiv');
+        save(fullfile(paths.EPI.dir,'AROMA/7_dataCSF.mat'),'CSFmask','CSFts','CSFavg','CSFderiv');
     end
     clear CSFts;
 %-------------------------------------------------------------------------%
@@ -958,11 +951,9 @@ if flags.EPI.Regressors == 1
     WMderiv = [0,diff(WMavg)];
     if configs.EPI.aCompCor == 1
         [WMpca,WMvar] = get_pca(WMts',configs.EPI.numCompPCA);
-        %save(fullfile(paths.EPI.dir,'AROMA/7_dataWM.mat'),'WMpca','WMvar','WMmask','WMts','WMavg','WMderiv');
-        save(fullfile(paths.EPI.dir,'GSreg_yes/7_dataWM.mat'),'WMpca','WMvar','WMmask','WMts','WMavg','WMderiv');
+        save(fullfile(paths.EPI.dir,'AROMA/7_dataWM.mat'),'WMpca','WMvar','WMmask','WMts','WMavg','WMderiv');
     else
-        %save(fullfile(paths.EPI.dir,'AROMA/7_dataWM.mat'),'WMmask','WMts','WMavg','WMderiv');
-        save(fullfile(paths.EPI.dir,'GSreg_yes/7_dataWM.mat'),'WMmask','WMts','WMavg','WMderiv');
+        save(fullfile(paths.EPI.dir,'AROMA/7_dataWM.mat'),'WMmask','WMts','WMavg','WMderiv');
     end
     clear WMts;
 %-------------------------------------------------------------------------%
@@ -972,8 +963,7 @@ if flags.EPI.Regressors == 1
     GSts = reshape(resting.vol(GSmask),[GSnumVoxels,numTimePoints]);
     GSavg = mean(GSts);
     GSderiv = [0,diff(GSavg)];
-    %save(fullfile(paths.EPI.dir,'AROMA/7_dataGS.mat'),'GSmask','GSts','GSavg','GSderiv');
-    save(fullfile(paths.EPI.dir,'GSreg_yes/7_dataGS.mat'),'GSmask','GSts','GSavg','GSderiv');
+    save(fullfile(paths.EPI.dir,'AROMA/7_dataGS.mat'),'GSmask','GSts','GSavg','GSderiv');
     clear GSts;
     
 %% -----------------------------------------------------------------------%
@@ -993,15 +983,13 @@ if flags.EPI.Regressors == 1
         disp('configs.EPI.GS not specified. Make sure its value is 1 or 0. Exiting...')
         return
     end
-    %save(fullfile(paths.EPI.dir,'AROMA/7_regressors.mat'),'regressors');
-    save(fullfile(paths.EPI.dir,'GSreg_yes/7_regressors.mat'),'regressors');
+    save(fullfile(paths.EPI.dir,'AROMA/7_regressors.mat'),'regressors');
 %-------------------------------------------------------------------------%
     % regress-out motion regressors  
     resting.vol(~GSmask)=0;
     resting.vol = apply_regressors(resting.vol,volBrain.vol,regressors);
     
-    %fileOut = fullfile(paths.EPI.dir,'AROMA/7_epi.nii.gz');
-    fileOut = fullfile(paths.EPI.dir,'GSreg_yes/7_epi.nii.gz');
+    fileOut = fullfile(paths.EPI.dir,'AROMA/7_epi.nii.gz');
     MRIwrite(resting,fileOut,'double');
 
     GSts_resid = reshape(resting.vol(GSmask),[GSnumVoxels,numTimePoints]); %time series of all voxels
@@ -1009,15 +997,10 @@ if flags.EPI.Regressors == 1
     WMts_resid = reshape(resting.vol(WMmask),[WMnumVoxels,numTimePoints]); %time series of WM voxels
     CSFts_resid = reshape(resting.vol(CSFmask),[CSFnumVoxels,numTimePoints]); %time series of CSF voxels
 
-%     save(fullfile(paths.EPI.dir,'AROMA/7_dataGSresid.mat'),'GSts_resid','GSmask');
-%     save(fullfile(paths.EPI.dir,'AROMA/7_dataGMresid.mat'),'GMts_resid','GMmask');
-%     save(fullfile(paths.EPI.dir,'AROMA/7_dataWMresid.mat'),'WMts_resid','WMmask');
-%     save(fullfile(paths.EPI.dir,'AROMA/7_dataCSFresid.mat'),'CSFts_resid','CSFmask');
-    
-    save(fullfile(paths.EPI.dir,'GSreg_yes/7_dataGSresid.mat'),'GSts_resid','GSmask');
-    save(fullfile(paths.EPI.dir,'GSreg_yes/7_dataGMresid.mat'),'GMts_resid','GMmask');
-    save(fullfile(paths.EPI.dir,'GSreg_yes/7_dataWMresid.mat'),'WMts_resid','WMmask');
-    save(fullfile(paths.EPI.dir,'GSreg_yes/7_dataCSFresid.mat'),'CSFts_resid','CSFmask');
+    save(fullfile(paths.EPI.dir,'AROMA/7_dataGSresid.mat'),'GSts_resid','GSmask');
+    save(fullfile(paths.EPI.dir,'AROMA/7_dataGMresid.mat'),'GMts_resid','GMmask');
+    save(fullfile(paths.EPI.dir,'AROMA/7_dataWMresid.mat'),'WMts_resid','WMmask');
+    save(fullfile(paths.EPI.dir,'AROMA/7_dataCSFresid.mat'),'CSFts_resid','CSFmask');
 
     clear GSts_resid GMts_resid WMts_resid CSFts_resid;
 
@@ -1033,10 +1016,8 @@ if flags.EPI.BandPass==1
     disp('8. Bandpass')
     disp('-----------')
 
-    %if exist(fullfile(paths.EPI.dir,'AROMA/7_dataGSresid.mat'),'file') 
-    %    load(fullfile(paths.EPI.dir,'AROMA/7_dataGSresid.mat'),'GSts_resid','GSmask');
-    if exist(fullfile(paths.EPI.dir,'GSreg_yes/7_dataGSresid.mat'),'file') 
-        load(fullfile(paths.EPI.dir,'GSreg_yes/7_dataGSresid.mat'),'GSts_resid','GSmask');
+    if exist(fullfile(paths.EPI.dir,'AROMA/7_dataGSresid.mat'),'file') 
+        load(fullfile(paths.EPI.dir,'AROMA/7_dataGSresid.mat'),'GSts_resid','GSmask');
     
         order = 1;
         f1 = (configs.EPI.fMin*2)*params.EPI.TR;
@@ -1044,12 +1025,10 @@ if flags.EPI.BandPass==1
         [tsf] = apply_butterworth_filter(GSts_resid',order,f1,f2);
         clear GSts_resid;
 
-        %fileIn = fullfile(paths.EPI.dir,'AROMA/7_epi.nii.gz');
-        fileIn = fullfile(paths.EPI.dir,'GSreg_yes/7_epi.nii.gz');
+        fileIn = fullfile(paths.EPI.dir,'AROMA/7_epi.nii.gz');
         resting = MRIread(fileIn);
         resting.vol(GSmask) = tsf';
-        %fileOut = fullfile(paths.EPI.dir,'AROMA/8_epi.nii.gz');
-        fileOut = fullfile(paths.EPI.dir,'GSreg_yes/8_epi.nii.gz');
+        fileOut = fullfile(paths.EPI.dir,'AROMA/8_epi.nii.gz');
         MRIwrite(resting,fileOut,'double');
         clear GSmask tsf resting;
     else
@@ -1064,15 +1043,13 @@ if flags.EPI.ROIs==1
     disp('11. ROIs')
     disp('--------')
     
-    %if exist(fullfile(paths.EPI.dir,'AROMA/8_epi.nii.gz'),'file')
-    if exist(fullfile(paths.EPI.dir,'GSreg_yes/8_epi.nii.gz'),'file')
+    if exist(fullfile(paths.EPI.dir,'AROMA/8_epi.nii.gz'),'file')
 
         volWM = MRIread(fullfile(paths.EPI.dir,'rT1_WM_mask_eroded.nii.gz')); volWM = volWM.vol;
         volCSF = MRIread(fullfile(paths.EPI.dir,'rT1_CSF_mask_eroded.nii.gz')); volCSF = volCSF.vol;
         volBrain = MRIread(fullfile(paths.EPI.dir,'rT1_brain_mask_FC.nii.gz')); volBrain = volBrain.vol;
 
-        %resting = MRIread(fullfile(paths.EPI.dir,'AROMA/8_epi.nii.gz'));
-        resting = MRIread(fullfile(paths.EPI.dir,'GSreg_yes/8_epi.nii.gz'));
+        resting = MRIread(fullfile(paths.EPI.dir,'AROMA/8_epi.nii.gz'));
             [sizeX,sizeY,sizeZ,numTimePoints] = size(resting.vol);
             
           for k=1:length(parcs.pdir)
@@ -1086,21 +1063,22 @@ if flags.EPI.ROIs==1
                         ROIs_numVoxels(i_roi,1)=nnz(parcGM==i_roi);
                     end
                     restingROIs = zeros(numROIs,numTimePoints);
+                    ROIs_numNans = nan(numROIs,numTimePoints);
                     for timePoint = 1:numTimePoints
                         aux = reshape(resting.vol(:,:,:,timePoint),[sizeX,sizeY,sizeZ]);
                         for ROI = 1:numROIs
                             voxelsROI = (parcGM==ROI);  
                             restingROIs(ROI,timePoint) = nanmean(aux(voxelsROI));
+                            ROIs_numNans(ROI,timePoint)=nnz(isnan(aux(voxelsROI)));
                         end
                         if mod(timePoint,50)==0
                             fprintf('%d out of %d\n',timePoint,numTimePoints);
                         end
                     end
-                    %fileOut=fullfile(paths.EPI.dir,strcat('AROMA/9_epi_',parcs.plabel(k).name,'_ROIs.mat'));
-                    fileOut=fullfile(paths.EPI.dir,strcat('GSreg_yes/9_epi_',parcs.plabel(k).name,'ROIs.mat'));
+                    fileOut=fullfile(paths.EPI.dir,strcat('AROMA/9_epi_',parcs.plabel(k).name,'_ROIs.mat'));
                     % ROIs_numVOxels is the number of voxels belonging to each node in a partition
                     % restingROIs is the average timeseries of each region
-                    save(fileOut,'restingROIs','ROIs_numVoxels');
+                    save(fileOut,'restingROIs','ROIs_numVoxels','ROIs_numNans');
                 end
            end      
     else
